Updated build of SFA.OPA.InterfaceTransform.Console to copy to location expected by powershell script

+++ src/OpaEarning/Build/Deploy/1718/sql/ddl/OPA.Transient.Rulebase.ddl.views.sql
@@ -3,6 +3,7 @@ if not exists(select schema_id from sys.schemas where name='Rulebase')
 GO
 
 if object_id('[Rulebase].[vw_AEC_Cases]','v') is not null
+
 	drop view [Rulebase].[vw_AEC_Cases]
 GO
 
@@ -12,7 +13,6 @@ AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
     ,[CaseData]
 FROM [Rulebase].[AEC_Cases]
 GO
-
 
 if object_id('[Rulebase].[vw_AEC_global]','v') is not null
 	drop view [Rulebase].[vw_AEC_global]
@@ -27,10 +27,11 @@ AS SELECT '' AS [Nothing]
 FROM [Rulebase].[AEC_global]
 GO
 
-
 if object_id('[Rulebase].[vw_AEC_LearningDelivery]','v') is not null
+
 	drop view [Rulebase].[vw_AEC_LearningDelivery]
 GO
+
 
 create view [Rulebase].[vw_AEC_LearningDelivery]
 AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
@@ -82,10 +83,11 @@ AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
 FROM [Rulebase].[AEC_LearningDelivery]
 GO
 
-
 if object_id('[Rulebase].[vw_AEC_LearningDelivery_Period]','v') is not null
+
 	drop view [Rulebase].[vw_AEC_LearningDelivery_Period]
 GO
+
 
 create view [Rulebase].[vw_AEC_LearningDelivery_Period]
 AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
@@ -122,10 +124,11 @@ AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
 FROM [Rulebase].[AEC_LearningDelivery_Period]
 GO
 
-
 if object_id('[Rulebase].[vw_AEC_LearningDelivery_PeriodisedValues]','v') is not null
+
 	drop view [Rulebase].[vw_AEC_LearningDelivery_PeriodisedValues]
 GO
+
 
 create view [Rulebase].[vw_AEC_LearningDelivery_PeriodisedValues]
 AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
@@ -147,10 +150,11 @@ AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
 FROM [Rulebase].[AEC_LearningDelivery_PeriodisedValues]
 GO
 
-
 if object_id('[Rulebase].[vw_AEC_LearningDelivery_PeriodisedTextValues]','v') is not null
+
 	drop view [Rulebase].[vw_AEC_LearningDelivery_PeriodisedTextValues]
 GO
+
 
 create view [Rulebase].[vw_AEC_LearningDelivery_PeriodisedTextValues]
 AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
@@ -172,10 +176,11 @@ AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
 FROM [Rulebase].[AEC_LearningDelivery_PeriodisedTextValues]
 GO
 
-
 if object_id('[Rulebase].[vw_AEC_HistoricEarningOutput]','v') is not null
+
 	drop view [Rulebase].[vw_AEC_HistoricEarningOutput]
 GO
+
 
 create view [Rulebase].[vw_AEC_HistoricEarningOutput]
 AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
@@ -207,8 +212,8 @@ AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
 FROM [Rulebase].[AEC_HistoricEarningOutput]
 GO
 
-
 if object_id('[Rulebase].[vw_AEC_ApprenticeshipPriceEpisode]','v') is not null
+
 	drop view [Rulebase].[vw_AEC_ApprenticeshipPriceEpisode]
 GO
 
@@ -216,44 +221,65 @@ create view [Rulebase].[vw_AEC_ApprenticeshipPriceEpisode]
 AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
     ,[LearnRefNumber]
     ,[PriceEpisodeIdentifier]
-    ,[EpisodeEffectiveTNPStartDate]
-    ,[EpisodeStartDate]
-    ,[PriceEpisodeActualEndDate]
-    ,[PriceEpisodeActualInstalments]
-    ,[PriceEpisodeAimSeqNumber]
-    ,[PriceEpisodeCappedRemainingTNPAmount]
-    ,[PriceEpisodeCompExemCode]
-    ,[PriceEpisodeCompleted]
-    ,[PriceEpisodeCompletionElement]
-    ,[PriceEpisodeContractType]
-    ,[PriceEpisodeCumulativePMRs]
-    ,[PriceEpisodeExpectedTotalMonthlyValue]
-    ,[PriceEpisodeFirstAdditionalPaymentThresholdDate]
-    ,[PriceEpisodeFundLineType]
-    ,[PriceEpisodeInstalmentValue]
-    ,[PriceEpisodePlannedEndDate]
-    ,[PriceEpisodePlannedInstalments]
-    ,[PriceEpisodePreviousEarnings]
-    ,[PriceEpisodePreviousEarningsSameProvider]
-    ,[PriceEpisodeRemainingAmountWithinUpperLimit]
-    ,[PriceEpisodeRemainingTNPAmount]
-    ,[PriceEpisodeSecondAdditionalPaymentThresholdDate]
-    ,[PriceEpisodeTotalEarnings]
-    ,[PriceEpisodeTotalPMRs]
-    ,[PriceEpisodeTotalTNPPrice]
-    ,[PriceEpisodeUpperBandLimit]
-    ,[PriceEpisodeUpperLimitAdjustment]
+    ,[TNP4]
     ,[TNP1]
+    ,[EpisodeStartDate]
     ,[TNP2]
     ,[TNP3]
-    ,[TNP4]
+    ,[PriceEpisodeUpperBandLimit]
+    ,[PriceEpisodePlannedEndDate]
+    ,[PriceEpisodeActualEndDate]
+    ,[PriceEpisodeTotalTNPPrice]
+    ,[PriceEpisodeUpperLimitAdjustment]
+    ,[PriceEpisodePlannedInstalments]
+    ,[PriceEpisodeActualInstalments]
+    ,[PriceEpisodeInstalmentsThisPeriod]
+    ,[PriceEpisodeCompletionElement]
+    ,[PriceEpisodePreviousEarnings]
+    ,[PriceEpisodeInstalmentValue]
+    ,[PriceEpisodeOnProgPayment]
+    ,[PriceEpisodeTotalEarnings]
+    ,[PriceEpisodeBalanceValue]
+    ,[PriceEpisodeBalancePayment]
+    ,[PriceEpisodeCompleted]
+    ,[PriceEpisodeCompletionPayment]
+    ,[PriceEpisodeRemainingTNPAmount]
+    ,[PriceEpisodeRemainingAmountWithinUpperLimit]
+    ,[PriceEpisodeCappedRemainingTNPAmount]
+    ,[PriceEpisodeExpectedTotalMonthlyValue]
+    ,[PriceEpisodeAimSeqNumber]
+    ,[PriceEpisodeFirstDisadvantagePayment]
+    ,[PriceEpisodeSecondDisadvantagePayment]
+    ,[PriceEpisodeApplic1618FrameworkUpliftBalancing]
+    ,[PriceEpisodeApplic1618FrameworkUpliftCompletionPayment]
+    ,[PriceEpisodeApplic1618FrameworkUpliftOnProgPayment]
+    ,[PriceEpisodeSecondProv1618Pay]
+    ,[PriceEpisodeFirstEmp1618Pay]
+    ,[PriceEpisodeSecondEmp1618Pay]
+    ,[PriceEpisodeFirstProv1618Pay]
+    ,[PriceEpisodeLSFCash]
+    ,[PriceEpisodeFundLineType]
+    ,[PriceEpisodeSFAContribPct]
+    ,[PriceEpisodeLevyNonPayInd]
+    ,[EpisodeEffectiveTNPStartDate]
+    ,[PriceEpisodeFirstAdditionalPaymentThresholdDate]
+    ,[PriceEpisodeSecondAdditionalPaymentThresholdDate]
+    ,[PriceEpisodeContractType]
+    ,[PriceEpisodePreviousEarningsSameProvider]
+    ,[PriceEpisodeTotProgFunding]
+    ,[PriceEpisodeProgFundIndMinCoInvest]
+    ,[PriceEpisodeProgFundIndMaxEmpCont]
+    ,[PriceEpisodeTotalPMRs]
+    ,[PriceEpisodeCumulativePMRs]
+    ,[PriceEpisodeCompExemCode]
 FROM [Rulebase].[AEC_ApprenticeshipPriceEpisode]
 GO
 
-
 if object_id('[Rulebase].[vw_AEC_ApprenticeshipPriceEpisode_Period]','v') is not null
+
 	drop view [Rulebase].[vw_AEC_ApprenticeshipPriceEpisode_Period]
 GO
+
 
 create view [Rulebase].[vw_AEC_ApprenticeshipPriceEpisode_Period]
 AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
@@ -283,8 +309,8 @@ AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
 FROM [Rulebase].[AEC_ApprenticeshipPriceEpisode_Period]
 GO
 
-
 if object_id('[Rulebase].[vw_AEC_ApprenticeshipPriceEpisode_PeriodisedValues]','v') is not null
+
 	drop view [Rulebase].[vw_AEC_ApprenticeshipPriceEpisode_PeriodisedValues]
 GO
 
@@ -307,46 +333,7 @@ AS SELECT (SELECT Ukprn FROM Valid.LearningProvider) AS [Ukprn]
     ,[Period_12]
 FROM [Rulebase].[AEC_ApprenticeshipPriceEpisode_PeriodisedValues]
 GO
-
-
-IF EXISTS (SELECT [object_id] FROM sys.views WHERE [name] = 'vw_AEC_EarningHistory' and [schema_id] = SCHEMA_ID('Rulebase')) 
-BEGIN 
-	DROP VIEW Rulebase.vw_AEC_EarningHistory 
-END
+IF EXISTS (SELECT [object_id] FROM sys.views WHERE [name] = 'vw_AEC_EarningHistory' and [schema_id] = SCHEMA_ID('Rulebase')) BEGIN DROP VIEW Rulebase.vw_AEC_EarningHistory END
 GO
-
-CREATE VIEW [Rulebase].[vw_AEC_EarningHistory] 
-AS SELECT 
-	[AppIdentifierOutput] [AppIdentifier], 
-	[AppProgCompletedInTheYearOutput][AppProgCompletedInTheYearInput], 
-	(SELECT[Name] FROM[Reference].[CollectionPeriods]) AS[CollectionReturnCode], 
-	'${YearOfCollection}' AS[CollectionYear], 
-	[HistoricDaysInYearOutput] [DaysInYear], 
-	[HistoricFworkCodeOutput] [FworkCode], 
-	[HistoricEffectiveTNPStartDateOutput] [HistoricEffectiveTNPStartDateInput], 
-	[HistoricLearner1618AtStartOutput] [HistoricLearner1618StartInput],
-	[HistoricTNP1Output] [HistoricTNP1Input], 
-	[HistoricTNP2Output] [HistoricTNP2Input],
-	[HistoricTNP3Output] [HistoricTNP3Input], 
-	[HistoricTNP4Output] [HistoricTNP4Input], 
-	[HistoricTotal1618UpliftPaymentsInTheYear] 
-	[HistoricTotal1618UpliftPaymentsInTheYearInput],
-	[HistoricVirtualTNP3EndofThisYearOutput] [HistoricVirtualTNP3EndOfTheYearInput],
-	[HistoricVirtualTNP4EndofThisYearOutput] [HistoricVirtualTNP4EndOfTheYearInput], 
-	1 AS[LatestInYear],
-	[LearnRefNumber],
-	[HistoricProgrammeStartDateIgnorePathwayOutput] [ProgrammeStartDateIgnorePathway], 
-	[HistoricProgrammeStartDateMatchPathwayOutput] [ProgrammeStartDateMatchPathway], 
-	[HistoricProgTypeOutput] [ProgType], 
-	[HistoricPwayCodeOutput] [PwayCode], 
-	[HistoricSTDCodeOutput] [STDCode], 
-	[HistoricTotalProgAimPaymentsInTheYear] [TotalProgAimPaymentsInTheYear], 
-	(SELECT Ukprn FROM Valid.LearningProvider) [UKPRN],
-	[HistoricULNOutput] [ULN],
-	[HistoricUptoEndDateOutput] [UptoEndDate], 
-	0.00 [BalancingProgAimPaymentsInTheYear], 
-	0.00 [CompletionProgAimPaymentsInTheYear], 
-	0.00 [OnProgProgAimPaymentsInTheYear]
-	
-FROM [Rulebase].[AEC_HistoricEarningOutput]
+CREATE VIEW [Rulebase].[vw_AEC_EarningHistory] AS SELECT [AppIdentifierOutput] [AppIdentifier], [AppProgCompletedInTheYearOutput][AppProgCompletedInTheYearInput], (SELECT[Name] FROM[Reference].[CollectionPeriods]) AS[CollectionReturnCode], '${YearOfCollection}' AS[CollectionYear], [HistoricDaysInYearOutput] [DaysInYear], [HistoricFworkCodeOutput] [FworkCode], [HistoricEffectiveTNPStartDateOutput] [HistoricEffectiveTNPStartDateInput], [HistoricLearner1618AtStartOutput] [HistoricLearner1618StartInput], [HistoricTNP1Output] [HistoricTNP1Input], [HistoricTNP2Output] [HistoricTNP2Input], [HistoricTNP3Output] [HistoricTNP3Input], [HistoricTNP4Output] [HistoricTNP4Input], [HistoricTotal1618UpliftPaymentsInTheYear] [HistoricTotal1618UpliftPaymentsInTheYearInput], [HistoricVirtualTNP3EndofThisYearOutput] [HistoricVirtualTNP3EndOfTheYearInput], [HistoricVirtualTNP4EndofThisYearOutput] [HistoricVirtualTNP4EndOfTheYearInput], 1 AS[LatestInYear], [LearnRefNumber], [HistoricProgrammeStartDateIgnorePathwayOutput] [ProgrammeStartDateIgnorePathway], [HistoricProgrammeStartDateMatchPathwayOutput] [ProgrammeStartDateMatchPathway], [HistoricProgTypeOutput] [ProgType], [HistoricPwayCodeOutput] [PwayCode], [HistoricSTDCodeOutput] [STDCode], [HistoricTotalProgAimPaymentsInTheYear] [TotalProgAimPaymentsInTheYear], (SELECT Ukprn FROM Valid.LearningProvider) [UKPRN], [HistoricULNOutput] [ULN], [HistoricUptoEndDateOutput] [UptoEndDate], 0.00 [BalancingProgAimPaymentsInTheYear], 0.00 [CompletionProgAimPaymentsInTheYear], 0.00 [OnProgProgAimPaymentsInTheYear]FROM [Rulebase].[AEC_HistoricEarningOutput]
 GO
